ai_worker/requirements.txt
requests>=2.31.0
pydantic>=2.7.0
python-dotenv>=1.0.1
tenacity>=8.2.3
